Enhance select_times to support repeated calls
The select_times method is now smart enough to support repeated calls,
allowing this to be used repeatedly to continue extracting ever-smaller
segments of a signal. Note that subsequent subsets must be fully
contained (i.e., you cannot request to select times that do not exist in
an already-subsetted signal).